refactor: app.js에서 home페이지 분리
