--- 몽실_김민정/week8/mission1/src/App.css
@@ -1,32 +1,0 @@
-.container {
-  margin: 30px;
-  width: 100vw;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.title {
-  margin-top: 30px;
-  margin-bottom: 100px;
-}
-
-.form {
-  width: 60%;
-}
-
-.input-wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  margin-bottom: 15px;
-}
-
-.list {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  margin-top: 50px;
-  width: 60%;
-}
